Natural scene recognition of remote sensing images/requirements.txt
scipy==1.4.1
scikit-learn==0.22.2.post1
scikit-image==0.16.2
pandas==1.0.3
numpy==1.18.3
Pillow==7.1.1
opencv-python
efficientnet_pytorch
timm